v1.1.0.34

+++ Game/Destruction/SCR_DestructionMultiPhaseComponent.c
@@ -1,5 +1,5 @@
 #define ENABLE_BASE_DESTRUCTION
-//------------------------------------------------------------------------------------------------
+
 enum SCR_EMaterialSoundTypeDebris
 {
 	NONE,
@@ -17,9 +17,8 @@ enum SCR_EMaterialSoundTypeDebris
 	SANDBAG,
 	WOOD_PLANK_SMALL,
 	WOOD_PLANK_LARGE
-};
+}
 
-//------------------------------------------------------------------------------------------------
 enum SCR_EMaterialSoundTypeBreak
 {
 	NONE,
@@ -38,14 +37,14 @@ enum SCR_EMaterialSoundTypeBreak
 	BREAK_TENT,
 	BREAK_WATERHYDRANT,
 	BREAK_WOOD_SOLID
-};
+}
 
-//------------------------------------------------------------------------------------------------
 [ComponentEditorProps(category: "GameScripted/Destruction", description: "Multi-Phase destruction component, for objects that go through several damage phases")]
-class SCR_DestructionMultiPhaseComponentClass: SCR_DestructionBaseComponentClass
+class SCR_DestructionMultiPhaseComponentClass: SCR_DestructionDamageManagerComponentClass
 {
 	[Attribute("1", UIWidgets.CheckBox, "If true, the object will be deleted after being damaged beyond the final damage phase", category: "Destruction Multi-Phase")]
 	bool m_bDeleteAfterFinalPhase;
+
 	[Attribute("", UIWidgets.Object, "List of the individual damage phases (sequential) above initial damage phase", category: "Destruction Multi-Phase")]
 	ref array<ref SCR_DamagePhaseData> m_aDamagePhases;
 	
@@ -54,11 +53,10 @@ class SCR_DestructionMultiPhaseComponentClass: SCR_DestructionBaseComponentClass
 	
 	[Attribute("0")]
 	float m_fMeshChangeDelay;
-};
+}
 
-//------------------------------------------------------------------------------------------------
 //! Multi-Phase destruction component, for objects that go through several destruction states
-class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
+class SCR_DestructionMultiPhaseComponent : SCR_DestructionDamageManagerComponent
 {
 #ifdef ENABLE_BASE_DESTRUCTION
 	private static int s_iFirstFreeDestructionMultiPhaseData = -1;
@@ -103,7 +101,7 @@ class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
 	}
 	
 	//------------------------------------------------------------------------------------------------
-	//! Returns the input damage phase data (null = initial)
+	//! \return the input damage phase data (null = initial)
 	SCR_DamagePhaseData GetDamagePhaseData(int damagePhase)
 	{
 		if (damagePhase <= 0 || damagePhase >= GetNumDamagePhases())
@@ -113,21 +111,21 @@ class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
 	}
 	
 	//------------------------------------------------------------------------------------------------
-	//! Returns whether the object should disable physics on destruction (before being deleted or changed)
+	//! \return whether the object should disable physics on destruction (before being deleted or changed)
 	override bool GetDisablePhysicsOnDestroy()
 	{
 		return false;
 	}
 	
 	//------------------------------------------------------------------------------------------------
-	//! Returns if the object is in the initial damage phase
+	//! \return if the object is in the initial damage phase
 	bool GetIsInInitialDamagePhase()
 	{
 		return GetDamagePhase() == 0;
 	}
 	
 	//------------------------------------------------------------------------------------------------
-	//! Returns which damage phase the object is currently in (0 = initial)
+	//! \return which damage phase the object is currently in (0 = initial)
 	int GetDamagePhase()
 	{
 		if (m_iDestructionMultiPhaseDataIndex == -1)
@@ -137,7 +135,7 @@ class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
 	}
 	
 	//------------------------------------------------------------------------------------------------
-	//! Returns the amount of damage phases (including initial)
+	//! \return the amount of damage phases (including initial)
 	int GetNumDamagePhases()
 	{
 		SCR_DestructionMultiPhaseComponentClass componentData = SCR_DestructionMultiPhaseComponentClass.Cast(GetComponentData(GetOwner()));
@@ -149,7 +147,11 @@ class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
 	}
 	
 	//------------------------------------------------------------------------------------------------
-	//! Returns the target damage stage for the input current health, current phase and damage
+	//! \return the target damage stage for the input current health, current phase and damage
+	//! \param[in] damagePhase
+	//! \param[in] health
+	//! \param[in] damage
+	//! \return
 	int CalculateDamagePhase(int damagePhase, float health, float damage)
 	{
 		if (damage < 0)
@@ -173,16 +175,17 @@ class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
 
 
 	//------------------------------------------------------------------------------------------------
+	//! \return
 	float GetDestructibleSize()
 	{
-		vector mins, maxs
+		vector mins, maxs;
 		GetOwner().GetBounds(mins, maxs);
 		return (maxs[0] - mins[0]) * (maxs[1] - mins[1]);
 	}
 	
 	//------------------------------------------------------------------------------------------------
-	// Update DamagePhase signal on entity that has SignalsManagerComponent
-	// Used for triggering looped sounds
+	//! Update DamagePhase signal on entity that has SignalsManagerComponent
+	//! Used for triggering looped sounds
 	protected void SetDamagePhaseSignal()
 	{
 		SignalsManagerComponent signalsManagerComponent = SignalsManagerComponent.Cast(GetOwner().FindComponent(SignalsManagerComponent));
@@ -232,14 +235,19 @@ class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
 	
 	//------------------------------------------------------------------------------------------------
 	//! Switches to the input damage phase (or deletes if past last phase)
+	//! \param[in] damagePhase
+	//! \param[in] delayMeshChange
 	void GoToDamagePhase(int damagePhase, bool delayMeshChange)
 	{
 		if (damagePhase >= GetNumDamagePhases()) // Going past final phase, so delete if we can
 		{
-			if (GetDestructionBaseData())
+			if (s_OnDestructibleDestroyed && GetDestructionBaseData())
 				s_OnDestructibleDestroyed.Invoke(this);
 			
+			SCR_DestructionMultiPhaseComponentClass componentData = SCR_DestructionMultiPhaseComponentClass.Cast(GetComponentData(GetOwner()));
+			if (componentData && componentData.m_bDeleteAfterFinalPhase)
 			DeleteDestructible();
+
 			return;
 		}
 		
@@ -265,6 +273,9 @@ class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
 	}
 	
 	//------------------------------------------------------------------------------------------------
+	//!
+	//! \param[in] damagePhaseData
+	//! \param[in] delayMeshChange
 	void ApplyDamagePhaseData(SCR_DamagePhaseData damagePhaseData, bool delayMeshChange)
 	{
 		ResourceName model;
@@ -346,6 +357,7 @@ class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
 		
 		// Set position
 		vector mat[4];
+		owner.GetTransform(mat);
 		
 		// Get position offset for slotted entities
 		vector mins;
@@ -375,6 +387,7 @@ class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
 	
 	//------------------------------------------------------------------------------------------------
 	//! Spawns objects that are meant to be created when the object is destroyed (particles, debris, etc)
+	//! \param[in] hitInfo
 	override void SpawnDestroyObjects(SCR_HitInfo hitInfo)
 	{
 		Physics ownerPhysics = GetOwner().GetPhysics();
@@ -384,7 +397,7 @@ class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
 		
 		if (GetIsInInitialDamagePhase()) // Initial phase
 		{
-			array<ref SCR_BaseSpawnable> destroySpawnObjects = SCR_DestructionBaseComponentClass.Cast(GetComponentData(GetOwner())).m_DestroySpawnObjects;
+			array<ref SCR_BaseSpawnable> destroySpawnObjects = SCR_DestructionDamageManagerComponentClass.Cast(GetComponentData(GetOwner())).m_DestroySpawnObjects;
 			int numSpawnOnDestroy = destroySpawnObjects.Count();
 			for (int i = 0; i < numSpawnOnDestroy; i++)
 			{
@@ -440,16 +453,24 @@ class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
 		if (!hitInfo)
 			return;
 		
-		if (hitInfo.m_TotalDestruction) // Total destruction, so just delete
+		SCR_DestructionMultiPhaseComponentClass componentData = SCR_DestructionMultiPhaseComponentClass.Cast(GetComponentData(GetOwner()));
+		if (!componentData)
+			return;
+		
+		if (hitInfo.m_TotalDestruction) // If total destruction, check if destructible should be deleted; if not, go to last damage phase
 		{
-			DeleteDestructible();
+			if (componentData.m_bDeleteAfterFinalPhase)
+				DeleteDestructible();
+			else
+				GoToDamagePhase(componentData.m_aDamagePhases.Count(), true);
 		}
 		else // Go to the stored target damage phase
 		{
-			if (hitInfo)
-				GetDestructionBaseData().DeleteHitInfo();
 			GoToDamagePhase(GetTargetDamagePhase(), true);
 		}
+		
+		if (hitInfo)
+			GetDestructionBaseData().DeleteHitInfo();
 	}
 	
 	//------------------------------------------------------------------------------------------------
@@ -505,7 +526,7 @@ class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
 	}
 	
 	//------------------------------------------------------------------------------------------------
-	protected override event bool OnRplLoad(ScriptBitReader reader)
+	protected override bool OnRplLoad(ScriptBitReader reader)
 	{
 		if (!super.OnRplLoad(reader))
 			return false;
@@ -533,11 +554,11 @@ class SCR_DestructionMultiPhaseComponent : SCR_DestructionBaseComponent
 		m_iOriginalHealth = GetHealth();
 	}
 #endif
-};
+}
 
-//------------------------------------------------------------------------------------------------
 class SCR_DamagePhaseTitle : BaseContainerCustomTitle
 {
+	//------------------------------------------------------------------------------------------------
 	override bool _WB_GetCustomTitle(BaseContainer source, out string title)
 	{
 		float health = 0;
@@ -545,17 +566,18 @@ class SCR_DamagePhaseTitle : BaseContainerCustomTitle
 		title = "Phase | HP: " + health.ToString();
 		return true;
 	}
-};
+}
 
-//------------------------------------------------------------------------------------------------
 //! Multi-Phase destruction phase
 [BaseContainerProps(), SCR_DamagePhaseTitle()]
 class SCR_DamagePhaseData
 {
 	[Attribute("100", UIWidgets.Slider, "Base health value for the damage phase. Upon switching to this phase, this value replaces the base health of the object", "0.01 100000 0.01")]
 	float m_fPhaseHealth;
+
 	[Attribute(ResourceName.Empty, UIWidgets.ResourcePickerThumbnail, "Model to use for the damage phase", "xob")]
 	ResourceName m_PhaseModel;
+
 	[Attribute("", UIWidgets.Object, "List of objects (particles, debris, etc) to spawn on destruction of the phase")]
 	ref array<ref SCR_BaseSpawnable> m_PhaseDestroySpawnObjects;
-};
+}
